command line remote resign
