+++ bin/isign
@@ -229,7 +229,7 @@ if __name__ == '__main__':
             # This could raise an ImportException but we'll let the exception bubble
             args.signer_class = import_class(args.signer_class_name)
 
-            signer_arguments = {}
+            args.signer_arguments = {}
             if args.signer_argument_pairs:
                 for pair in args.signer_argument_pairs:
                     if '=' not in pair:
